54 UARTE: Add support for Address bit and configurable data size
Note: The FIFO backend now moves 16 bits data per frame instead of
8bits.
So users must recompile their devices if they connect them
between each other with this UART backend.

Note: In Rx, logging the data to file now happens before doing a
possible Rx callback

+++ src/HW_models/NHW_UART.h
@@ -78,7 +78,7 @@ extern NRF_UARTE_Type NRF_UARTE_regs[];
 extern NRF_UART_Type *NRF_UART_regs[];
 #endif
 
-typedef void (*uart_rtxb_cb_f)(uint inst, uint8_t *data);
+typedef void (*uart_rtxb_cb_f)(uint inst, uint16_t *data);
 /*
  * (Test interface) Register a callback which will be called
  * each time a byte is transmitted or received.
